Rename the settings module.
