+++ src/backend/aspen/api/main.py
@@ -4,10 +4,10 @@ from fastapi import Depends, FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
 from aspen.api.auth import get_auth_user
-from aspen.api.config.config import get_settings
 from aspen.api.deps import set_db
 from aspen.api.error.http_exceptions import AspenException, exception_handler
 from aspen.api.middleware.session import SessionMiddleware
+from aspen.api.settings import get_settings
 from aspen.api.views import auth, health, users
